SimpleTensorNetworks

This is the documentation for SimpleTensorNetworks.

API

Index

SimpleTensorNetworks.Index — Type

Simple index type

Indices with the same name and the same dimension are considered to be the same index.

SimpleTensorNetworks.isunique — Method

Returns if a given vector/set of indices contains only unique indices

IndexedArray

SimpleTensorNetworks.IndexedArray — Type

Simple Array with indices

Base.length — Method

Get the length of an IndexedArray

Base.size — Method

Get the size of an IndexedArray

SimpleTensorNetworks.data — Method

Get the underlying data of an IndexedArray

SimpleTensorNetworks.hascommondindices — Method

Return if two IndexedArrays have common indices

SimpleTensorNetworks.indices — Method

Get the indices of an IndexedArray

SimpleTensorNetworks.permute — Method

Permute the indices of an IndexedArray

indices: The new indices (list, tuple, etc.)

contraction

SimpleTensorNetworks.contract — Method

Contract two IndexedArrays over the shared indices.

SimpleTensorNetwork

SimpleTensorNetworks.TensorNetwork — Type

Generic tensor network data structure

Graphs.is_cyclic — Method

Return if a tensor network tn has a cycle. If it has not a cycle, tn is a tree tensor network.

SimpleTensorNetworks._contract_subtree — Method

Contract all the tensors in a subtree of a tensor network tn and return the result.

The subtree is defined by a vertex v and its parent vertex parent_v. Note that parent_v is not included in the subtree.

SimpleTensorNetworks.complete_contraction — Method

Contract all the tensors in a tensor network tn and return the result.

This function works only for tree tensor networks, i.e., is_cyclic(tn) == false.

root_vertex: The vertex to start the contraction. The default is 1.